refactor: rename identityPlugin.js to identity-plugin.js
